requirements.txt
# Core dependencies
pefile>=2023.2.7
loguru>=0.7.2
rich>=13.6.0
keystone-engine>=0.9.2
capstone>=4.0.2
typing-extensions>=4.8.0
psutil>=5.9.5

# Development dependencies (optional)
pytest>=7.4.2
black>=23.9.1
pylint>=3.0.1
mypy>=1.5.1

# Documentation dependencies (optional)
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0 